requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0

langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langgraph==0.2.35

python-dotenv==1.0.1
httpx==0.27.2
tavily-python==0.5.0

psycopg[binary]==3.2.13
pgvector==0.3.2
sqlalchemy==2.0.35